requirements.txt
firebase-admin>=5.0.0
python-dotenv>=1.0.0
requests>=2.31.0
pytz>=2024.1
google-generativeai>=0.3.0
apscheduler>=3.10.0
flask>=2.0.0
gunicorn>=20.0.0
python-dateutil
